# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalyticsEventRequest": {
        "properties": {
          "name": {
            "enum": [
              "onboarding_saved",
              "baseline_assessment_submitted",
              "coach_rewrite_requested",
              "role_play_turn_submitted",
              "privacy_export_requested",
              "privacy_deletion_requested",
              "moderation_report_submitted",
              "accessibility_preference_changed"
            ],
            "title": "Name",
            "type": "string"
          },
          "properties": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalyticsProperties"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "enum": [
              "web",
              "mobile",
              "api"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "name",
          "source"
        ],
        "title": "AnalyticsEventRequest",
        "type": "object"
      },
      "AnalyticsEventResponse": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "eventId": {
            "title": "Eventid",
            "type": "string"
          },
          "receivedAt": {
            "format": "date-time",
            "title": "Receivedat",
            "type": "string"
          }
        },
        "required": [
          "accepted",
          "eventId",
          "receivedAt"
        ],
        "title": "AnalyticsEventResponse",
        "type": "object"
      },
...
      "BaselineAssessmentAnswer": {
        "properties": {
          "questionId": {
            "title": "Questionid",
            "type": "string"
          },
          "value": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Value",
            "type": "integer"
          }
        },
        "required": [
          "questionId",
          "value"
        ],
        "title": "BaselineAssessmentAnswer",
        "type": "object"
      },
      "BaselineAssessmentQuestion": {
        "properties": {
          "highLabel": {
            "title": "Highlabel",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lowLabel": {
            "title": "Lowlabel",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "id",
          "prompt",
          "lowLabel",
          "highLabel"
        ],
        "title": "BaselineAssessmentQuestion",
        "type": "object"
      },
      "BaselineAssessmentRequest": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/BaselineAssessmentAnswer"
            },
            "title": "Answers",
            "type": "array"
          }
        },
        "required": [
          "answers"
        ],
        "title": "BaselineAssessmentRequest",
        "type": "object"
      },
      "BaselineAssessmentResponse": {
        "properties": {
          "recommendedLessonIds": {
            "items": {
              "type": "string"
            },
            "title": "Recommendedlessonids",
            "type": "array"
          },
          "score": {
            "$ref": "#/components/schemas/FeedbackScore"
          },
          "style": {
            "enum": [
              "accommodating",
              "direct",
              "avoidant",
              "balanced",
              "intense"
            ],
            "title": "Style",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "style",
          "summary",
          "score",
          "recommendedLessonIds"
        ],
        "title": "BaselineAssessmentResponse",
        "type": "object"
      },
      "CoachRewriteRequest": {
        "properties": {
          "context": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "goal": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          },
          "inputText": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Inputtext",
            "type": "string"
          },
          "relationship": {
            "enum": [
              "coworker",
              "manager",
              "friend",
              "partner",
              "family",
              "service-provider",
              "other"
            ],
            "title": "Relationship",
            "type": "string"
          },
          "tone": {
            "enum": [
              "warm-direct",
              "firm-boundary",
              "repair",
              "curious"
            ],
            "title": "Tone",
            "type": "string"
          }
        },
        "required": [
          "inputText",
          "relationship",
          "tone",
          "goal"
        ],
        "title": "CoachRewriteRequest",
        "type": "object"
      },
      "CoachRewriteResponse": {
        "properties": {
          "assertiveText": {
            "title": "Assertivetext",
            "type": "string"
          },
          "coachingNotes": {
            "items": {
              "$ref": "#/components/schemas/CoachingNote"
            },
            "title": "Coachingnotes",
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "feedbackScore": {
            "$ref": "#/components/schemas/FeedbackScore"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "outputSafetyFlags": {
            "items": {
              "type": "string"
            },
            "title": "Outputsafetyflags",
            "type": "array"
          },
          "safetyFlags": {
            "items": {
              "type": "string"
            },
            "title": "Safetyflags",
            "type": "array"
          },
          "suggestedPractice": {
            "items": {
              "type": "string"
            },
            "title": "Suggestedpractice",
            "type": "array"
          }
        },
        "required": [
          "id",
          "assertiveText",
          "coachingNotes",
          "feedbackScore",
          "safetyFlags",
          "outputSafetyFlags",
          "suggestedPractice",
          "createdAt"
        ],
        "title": "CoachRewriteResponse",
        "type": "object"
      },
      "CoachingNote": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "detail"
        ],
        "title": "CoachingNote",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModerationReportRequest": {
        "properties": {
          "details": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "reason": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "subjectId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectid"
          },
          "subjectType": {
            "enum": [
              "coach_message",
              "practice_attempt",
              "scenario",
              "other"
            ],
            "title": "Subjecttype",
            "type": "string"
          }
        },
        "required": [
          "subjectType",
          "reason"
        ],
        "title": "ModerationReportRequest",
        "type": "object"
      },
      "ModerationReportResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "open",
              "triaged",
              "resolved",
              "dismissed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "createdAt"
        ],
        "title": "ModerationReportResponse",
        "type": "object"
      },
      "ModerationReportUpdate": {
        "properties": {
          "reviewerNotes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewernotes"
          },
          "status": {
            "enum": [
              "open",
              "triaged",
              "resolved",
              "dismissed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ModerationReportUpdate",
        "type": "object"
      },
...
      "PracticeAttemptRequest": {
        "properties": {
          "draftText": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Drafttext",
            "type": "string"
          },
          "scenarioId": {
            "title": "Scenarioid",
            "type": "string"
          }
        },
        "required": [
          "scenarioId",
          "draftText"
        ],
        "title": "PracticeAttemptRequest",
        "type": "object"
      },
      "PracticeAttemptResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/CoachingNote"
            },
            "title": "Notes",
            "type": "array"
          },
          "score": {
            "$ref": "#/components/schemas/FeedbackScore"
          }
        },
        "required": [
          "id",
          "score",
          "notes",
          "createdAt"
        ],
        "title": "PracticeAttemptResponse",
        "type": "object"
      },
...
      "PrivacyDeletionRequest": {
        "properties": {
          "deleteAccount": {
            "title": "Deleteaccount",
            "type": "boolean"
          },
          "deleteAssessment": {
            "title": "Deleteassessment",
            "type": "boolean"
          },
          "deletePracticeHistory": {
            "title": "Deletepracticehistory",
            "type": "boolean"
          }
        },
        "required": [
          "deletePracticeHistory",
          "deleteAssessment",
          "deleteAccount"
        ],
        "title": "PrivacyDeletionRequest",
        "type": "object"
      },
      "PrivacyDeletionResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "estimatedCompletion": {
            "format": "date-time",
            "title": "Estimatedcompletion",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "completed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "createdAt",
          "estimatedCompletion"
        ],
        "title": "PrivacyDeletionResponse",
        "type": "object"
      },
      "PrivacyExportResponse": {
        "properties": {
          "downloadUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Downloadurl"
          },
          "format": {
            "const": "json",
            "title": "Format",
            "type": "string"
          },
          "generatedAt": {
            "format": "date-time",
            "title": "Generatedat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "includes": {
            "items": {
              "type": "string"
            },
            "title": "Includes",
            "type": "array"
          }
        },
        "required": [
          "id",
          "generatedAt",
          "format",
          "includes"
        ],
        "title": "PrivacyExportResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Coaching, practice, progress, and moderation API for SpeakAble.",
    "title": "SpeakAble API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/moderation/reports": {
      "get": {
        "operationId": "list_reports_v1_admin_moderation_reports_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModerationReportResponse"
                  },
                  "title": "Response List Reports V1 Admin Moderation Reports Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reports",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/moderation/reports/{report_id}": {
      "patch": {
        "operationId": "update_report_v1_admin_moderation_reports__report_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerationReportUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerationReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Report",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/analytics/events": {
      "post": {
        "operationId": "track_event_v1_analytics_events_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsEventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Track Event",
        "tags": [
          "analytics"
        ]
      }
    },
    "/v1/assessment/baseline": {
      "post": {
        "operationId": "submit_baseline_v1_assessment_baseline_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineAssessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineAssessmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Baseline",
        "tags": [
          "assessment"
        ]
      }
    },
    "/v1/assessment/baseline/questions": {
      "get": {
        "operationId": "list_questions_v1_assessment_baseline_questions_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BaselineAssessmentQuestion"
                  },
                  "title": "Response List Questions V1 Assessment Baseline Questions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Questions",
        "tags": [
          "assessment"
        ]
      }
    },
    "/v1/coach/rewrite": {
      "post": {
        "operationId": "create_rewrite_v1_coach_rewrite_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoachRewriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoachRewriteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rewrite",
        "tags": [
          "coach"
        ]
      }
    },
...
    "/v1/moderation/reports": {
      "post": {
        "operationId": "create_report_v1_moderation_reports_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerationReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerationReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Report",
        "tags": [
          "moderation"
        ]
      }
    },
...
    "/v1/practice-attempts": {
      "post": {
        "operationId": "create_practice_attempt_v1_practice_attempts_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeAttemptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeAttemptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Practice Attempt",
        "tags": [
          "coach"
        ]
      }
    },
    "/v1/privacy/delete": {
      "post": {
        "operationId": "request_deletion_v1_privacy_delete_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivacyDeletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyDeletionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Deletion",
        "tags": [
          "privacy"
        ]
      }
    },
    "/v1/privacy/export": {
      "post": {
        "operationId": "export_privacy_data_v1_privacy_export_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Privacy Data",
        "tags": [
          "privacy"
        ]
      }
    },
...
  }
}